PFDR-255 - Add Bentley Audio as a top level content type
This is not an ideal solution, but a practical one. From the
repository's point of view, audio from Bentley is not the same as plain
audio. However, it behaves the same as audio with local metadata. This
commonality and variance give us repetition and subtle difference in the
code, which is subject to refactoring. The best way to do that is not
clear now (how to compose optional/required behavioral elements and
reduce duplication in CLI and bagger test coverage) and I don't want to
block usage on a refactoring expedition that potentially spans three or
four repos.