PFDR-260 - Move back to basic rsync flags
The rsync command mentioned in the man page around --no-perms and
--chmod does work as expected, provided that the umask is set 0002 on
the server. I figured out why my setting for that was failing and that
trying to get the setgid bit with --perms was more hassle. It came down
to the pam_umask settings needing to be for sshd, not login.