PFDR-269 - Add batch helper scripts
These are a couple of utilities to help with bagging and uploading
batches, each with some logging. Putting chipmunk-client/bin on the PATH
and setting CHIPMUNK_LOGS to a writable path for logging is the easiest
way to use these. Alternatively, setting CHIPMUNK_HOME to the client
directory and using an explicit path to batch-bag or batch-upload will
also work.