PFDR-281 - Fix up test suite and add Dockerfile
There were some issues in the test suite after PFDR-275, where we added
intentional output at some stages of the uploader. There was also quite
a bit of extraneous output while running the suite, and this change
silences all of that.

There is now a Dockerfile and docker-compose.yml to make it easy to run
the suite in a container, since there are some Linux dependencies in the
tests. Using docker-compose run dev will run everything.