Set up mocha and cucumber
This is the first swipe at a full test harness. It's a little bit wacky
still, because we aren't using the WDIO test runner and its sync mode
just yet. This makes selectors and async calls to the client/DOM much
more cumbersome. The next step is to see if we can get sync wrapper
working, or integrate everything under the WDIO runner.