The UI driver is completely hooked into the UI
There are some obvious smells, and we should probably use a component
instead of having the UI driver know about specific html elements and
classes, but it is communicating effectively with the UI.

The UI is currently lying about whether it's packaging anything because
right now we don't know how to package anything, so we'll also want a
way for the UI to hook into our code. That is, instead of generating a
list of packages from a list of artifacts, the UI should send a list of
artifact locations (and a content type) to our library, and it should
receive a list of packages in return.

Third, our setup module continues not to know how to set up (or tear
down) a digital artifact.