Begin driving out domain objects, events, and services
After some considerable trips around the loop, we are now clarifying the
more specific responsibilities and data needed to make the interactor
pattern work. The domain objects to represent an artifact in its
various states, and the events related to packaging have emerged, as
well as the collaborating services to provide the packaging use case.