Implement Packager lifecycle
- Set up packaging as sequential awaits with one catch
- Pass in a listener so packager can notify on events
- Extract PackageEvents from test to src/PackagingListener

I am extremely satisifed with the workflow method of package() now.
Because the Packager holds everything it needs to operate, these can be
no-arg methods called in sequence. With the try-catch sugar that
async/await give over promise resolution/rejection, the fall-out case
perfectly reveals the operation.

I also really like the separation of the lifecycle methods from the
interrogations or external commands. This specification/implementation
split within the class allows all of the strict enforcement and error
translation to happen in methods addressed only to self, while all of
the collaboration happens in unconditional methods that just answer the
question or conduct the action without handling any errors. Rephrased,
methods like sourceReadable add packaging semantics over the filesystem
operation, but do not imply the consequences to the workflow. That is,
they are queries and commands that make no judgment on what to do with
the result -- they are objectively simpler, and so, more stable.

Also, I like carrots.