Make sure that the "garbage" test is failing for the right reason
The type check was not throwing, but the test was still passing because
the check on the source path was throwing and we were only verifying
that we don't make the bag.

There is a question of how much overspecification "does not bag" is, but
I think that will go away when there is some differentiation of the
cause for failure. I think failed() contract will be to send a packaging
error that refers to the artifact, rather than simply sending the
artifact. If that's the case, the expectations become that failed was
called with the right type of error/message and that it's for the
artifact, and the "does not attempt to bag" goes away here.