Revert to lockfile v1 / npm6, fix Packager import
There is some kind of conflict with lzma-native/node-gyp and npm 7 on
Linux... Whatever... Node 16 will release soon and this will probably be
surfaced and fixed shortly thereafter. For now, use npm 6 to package.

Also, fix an import in the interactors file, which was passing on Mac
despite being capitalized.