Detect SEVERE browser errors and fail scenarios
This adds a checkErrors method to the UI driver and a hook to run it on
UI-tagged scenarios before stopping the app. It reaches into the browser
object and gets its logs, throwing if any SEVERE items are logged. This
corresponds to console.error entries from the browser. We may want to
add more subtlety later but, for now, this catches things that leak all
the way back, like if there is a problem running a docker command.