Add new wireframe for main layout
- Add Foundation
- Add jQuery
- Add DataTables
- Add the CSS grid for main UI sections

This patch is pretty straightforward, but it took considerable effort to
derive the formula for getting the scripts and styles included at the
right places and to make jQuery available to the two other libraries
that use it. I had the activation in a svelte:head, but it seems like
onMount is a little more appropriate and clear about the timing.

The window and document are both definitely available in svelte.js, so
that's nice to confirm for understanding how Rollup, Svelte, and
Electron all fit together in a more detailed way.