Check in messy prototype of interactions
This is a complete mess, but it's starting to illustrate an idea of
how the interactions might play out. This should serve as a spike; that
is, be thrown out once we sketch out the target.

Basically, the interactions with the main/renderer processes and how
they connect to state storage was totally unknown but there are some
ideas and options in here, somewhere.